[BE]'Feat': "progress 구현 db 저장확인 완료"

+++ server_9dokme/src/main/java/com/example/server_9dokme/rent/entity/Rent.java
@@ -26,7 +26,7 @@ public class Rent extends BaseEntity {
     @DateTimeFormat(pattern = "yyyy.MM.dd HH:mm:ss")
     private LocalDateTime readAt;
 
-    private Long progress;
+    private float progress;
 
     private int lastPage;
